+++ src/main/kotlin/io/unthrottled/doki/notification/UpdateNotification.kt
@@ -8,7 +8,9 @@ import com.intellij.notification.NotificationGroup
 import com.intellij.notification.NotificationListener
 import com.intellij.notification.NotificationType
 import com.intellij.notification.impl.NotificationsManagerImpl
+import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.project.ProjectManager
 import com.intellij.openapi.ui.popup.Balloon
 import com.intellij.openapi.wm.WindowManager
 import com.intellij.ui.BalloonLayoutData
@@ -63,15 +65,28 @@ object UpdateNotification {
     @Nls(capitalization = Nls.Capitalization.Sentence) title: String = "",
     @Nls(capitalization = Nls.Capitalization.Sentence) content: String,
     project: Project? = null,
-    listener: NotificationListener? = defaultListener
+    listener: NotificationListener? = defaultListener,
+    actions: List<AnAction> = emptyList(),
   ) {
     val notification = notificationGroup.createNotification(
       title,
       content,
       listener = listener
     ).setIcon(NOTIFICATION_ICON)
+    notification.addActions(actions)
     notification.isImportant = true
     notification.notify(project)
+  }
+
+  fun showNotificationAcrossProjects(
+    @Nls(capitalization = Nls.Capitalization.Sentence) title: String = "",
+    @Nls(capitalization = Nls.Capitalization.Sentence) content: String,
+    listener: NotificationListener? = defaultListener,
+    actions: List<(Project?) -> AnAction> = emptyList()
+  ) {
+    ProjectManager.getInstance().openProjects.forEach { project ->
+      showDokiNotification(title, content, project, listener, actions.map { it(project) })
+    }
   }
 
   fun sendMessage(