+++ changelog/RELEASE-NOTES.md
@@ -6,5 +6,13 @@
 - Many small changes to every theme's look and feel [see issue for more details](https://github.com/doki-theme/doki-theme-jetbrains/issues/306)
 - Restored background image functionality.
   - You can now have your own background, and your Waifu sticker now!
-  - Stay tuned for official Doki-Theme background images!!
   - ![Ishtar's Background](https://raw.githubusercontent.com/doki-theme/doki-theme-jetbrains/master/assets/doki-theme-background-ishtar-dark.png)
+- All the plugin's themes' frame backgrounds are now available at your convenience!
+  - Use the `Enable Wallpaper` action or enable it in the Settings menu!
+  - Warning! The installed wallpapers will remain on your IDE after uninstalling the plugin.
+    You still can edit/remove the background using the "Set Background Image" action.
+    Turning off the setting/action will do the same thing.
+  - See the [pull request for more information](https://github.com/doki-theme/doki-theme-jetbrains/pull/312)
+  - ![Megumin](https://user-images.githubusercontent.com/15972415/104955009-0ee4c080-598f-11eb-9948-aaa8c5ef18d8.png)
+
+